\name{obesity}
\alias{obesity}
\docType{data}
\title{
Body mass index
}
\description{
A cross-sectional dataset of real microbiome
}
\usage{data(obesity)}
\format{
phyloseq
}

\references{
McMurdie PJ and Holmes S. phyloseq: An R package for reproducible interactive analysis and graphics of microbiome census data. PLoS ONE. 2013;8(4):e61217

Escobar JS, et al. The gut microbiota of Colombians differs from that of Americans, Europeans and Asians. BMC Microbiology. 2014;14(1):311.

}

\examples{
library(phyloseq)

data(obesity)
obesity

}
